Added set_file function to change which file is being logged to after original setup.
This is so that all threads can be included in logging, even when you don't know what the log file should be called yet. So you can call log() as early as you want in your program, and then call set_file when you know what you want to rename the log file to.
For example in analysis_subprocess, we would like the log files to be named after the analysis routines of each process, but we don't know what the routine is until after some inititialisation code which starts some threads. These threads won't be logged unless we call log() before starting them.