Modified invoke_in_main to not "leak" memory
I put "leak" in quotes because it is yet unclear whether it actually does leak memory. What we do know is that the memory usage of a program using the in_main function does grow, but periodically it drops significantly. Longer testing must be done before we can tell whether we actually have a memory leak.

However, I have discovered that the Qt documentation states that an event passed to QCoreApplication.postEvent() should not be accessed after it is posted. We were accessing it to wait for the event to be processed. When we don't access it after calling QCoreApplication.postEvent() (and instead pass in a queue at event instantiation, which we then wait on), the memory usage does not grow nearly as rapidly and from preliminary observations, oscialltes around an average usage.

Further testing will ensue.