1.1.0-dev: labonfig now looks in labscript install directory first before hardcoded paths.
Fixed a bug in h5_lock where it would start zlock server in the current
working directory. On Windows, this would prevent deletion of that directory
later, as zlock would still be running and Windows won't let you delete folders
that are the working directory of a process. It now starts in %temp%.