1.1.0-dev: Fixed a regression introduced where the winshell module relied on the labscript suite being installed.
All I needed was a unique string identifying the installation. Now, if there is
no installation, this is set to <not_installed>, rather than throwing an
exception.

This is just used for setting the AppIDs. It is useful for them to
be unique across installations, as otherwise windows can be slow to update
some sort of cache and taskbar behaviour can be wrong until it does so.