feature: Change to re-lay out widgets immediately, instead of after a timeout.
This prevents an error when the palette is destroyed before the timeout,
and it improves responsiveness and doesn't seem to be a performance problem.

Remove HLine frame from ToolPalette widget - it drew some silly tiny black lines
in the widget that I didn't notice at first. I wanted to get rid of the frame
earlier but it seemed to prevent resizes. Then I realised that the frame
contributed 2 pixels to the size of the widget, and so a pixel maths in
the resize event just needed to be adjusted by 2 pixels when the frame was
removed.