Python3: Make qtwidgets require qtutils 2.1.0
Make properties be able to extract JSON properties regardless of whether they
are stored in the file as fixed length bytestrings or as variable length
unicode strings, and regardless of whether the device name is stored as either
type of string.

This frees the way for future labscript to save variable length
unicode strings, with all code still able to read both type of strings.