Python3: Manually inspected 2to3 results to determine any code potentially in need of attention.
More kosher version check in excepthook
list(dict) fix in excepthook

Explicit use of standard library configparser in labconfig in python2 so that
we don't depend on the backported configparser module (we can decide to do
so later if we want to start using Python3 features in the configparser
module while retaining backward compatbility but I doubt we will need this
before Python 2 compat is dropped anyway).

No more issues highlighted by inspection of 2to3 results.