winlauncher: Move functionality from the installer into winshell module for consistent app descriptions (app descriptions in the taskbar vs start menu differed previously), and to allow apps to know at run time what the command to launch them is, since they need to set it as their relaunch command.
Add launcher that runs apps in a conda environment, and that if run with pythonw
runs apps in python.exe without a console window. This solves a number of issues
relating to pythonw and makes zprocess's output redirection more comprehensive.
For example, C/cython extensions or non-python subprocess output cannot be
redirected to output boxes when running under pythonw. Starting the launcher
with pythonw.exe and then having it run the target with python.exe allows for
no console window to be visible, even for a moment.

Using this launcher instead of cwp.py from the menuinst project is also required
because our apps need more environment variables pertaining to conda than are
set by cwp.py. Further, cwp.py unconditionally sets the working directory
to the user's documents directory, whereas we have configured out own working
directories that we would prefer our apps to run in.