h5_lock_better_error: Fix the bug that motivated the better error message.
Using the better error message, the bug was located and fixed.
The problem was that as of the serialised_attributes changes, the
labscript_utils.properties module imported h5py but not h5_lock, and so code
that imported labscript_utils.properties prior to h5_lock would cause the error
to be raised.

Rather than have labscript_utils.properties import h5_lock, it now lazy-imports
h5py within the functions that need it, so that the import order no longer
matters.