check_version: Compatibility with importlib_metadata >= 0.21
importlib_metadata's interface changed in this version in a way that broke
our functions. They have now been updated to work with newer versions of
importlib_metadata, whilst remaining compatible with older versions for now
(at the very least we need to keep compatibility until new importlib_metadata is
in the anaconda repositories).