get versions using metadata
In setup.py, use setuptools_scm to set the project version
from git tags. If not on a tagged release, it will create
a development version number in a format like:

2.15.1.dev34+g96c8c11.d20200430

Extend labscript_utils.versions.get_version() to first look for git
repository metadata to determine the version of a package, in the case
that it is being imported from a git repository.

Do not hard-code `__version__` - instead determine own version using
a call to get_version().

This means that for a git repository, the single source of truth for the
version is the repository metadata. For an installed package, the single
source of truth is the installation metadata. So long as packages use
the default version format for `setuptools_scm`, the version from a git
repo and then the subsequent installed package should be identical.