Remove transitional code for migrating log files.
This was causing issues due to trying to import blacs under the name
BLACS, which succeeds on Windows due to case insensitivity but is a
double import picked up by double import denier.

Migrating log files is not so important since nobody will be upgrading
in-place to the new packages after the move to github.