Move winlauncher to be a top-level module
This way winlauncher can be imported without importing anything
labscript_utils related. This will facilitate the creation of
entry_points for other applications, that import winlauncher in order
to run the application with it.

This commit is the rename of winlauncher, in order to make the diff
meaningful, modifying it to function as an importable module will be done
in subsequent commits.