Allow winlauncher to be an entry_point for apps
Make it so that another package may define entry points in its
setup() as (using runmanager as an example):

```
entry_points={
        'console_scripts': ['runmanager = winlauncher:main'],
        'gui_scripts': ['runmanagerw = winlauncher:main'],
    },
```

and that this will have winlauncher run the `__main__.py` of the module
in question, after activating the conda environment associated with the
Python executable that was used to install the entry point.

The convention is that the entry points should be named `<modulename>`
and `<modulenamew>` (for the non-console version). `winlauncher` sees
this as its sys.argv[0] and therefore knows which module to run.

Conda environment setup is now inspected based on the location of the
Python executable, so is no longer passed in on the command line.

This setup makes there be a runmanager.exe and runmanagerw.exe within
the Scripts directory of the Python installation, that can be used to
start runmanager directly so long as one has the full path to it.

Yet to do is update winshell to set relaunch commands correctly in light
of this change.

Conda env changes are untested for now.