Remove unused modules.
clean_repo_hook is Python2 and mercurial specific. (Python2 because the
presence of .pyc files within a `__pycache__` directory as is done in
PY3 does not automatically allow the module to be imported when the .py
file is absent). Although we might want to bring something like it back,
it's not super necessary now that we use neither.

Winshell and Winlauncher are superseded by the desktop-app project.

numpy_dtpye_workaround is no longer needed since numpy 1.15, and there
are PRs removing it from the labscript projects still importing it.

This will require a major version bump in labscript_utils.

Code using check_version will break on this, they should either update
to allow the newer version or stop using check_version on
labscript_utils since packaging tools will handle dependencies now.