Move copying of .pth file into setup.py
Use a custom develop command to copy the .pth file into site-packages
when `python setup.py develop` is run - this is what pip does under the
hood for `pip install -e`.

No longer need desktop-app to locate site-packages, as the setuptools
command class knows it as `self.install_dir`.

The file is not removed by `pip uninstall`, but due to the `try:
except:`s in the `.pth` file, leaving it behind is harmless.

It is removed if the user runs whatever `setup.py` command undoies