Include labscript-suite.pth in editable wheels
This ensures user directories are added to the Python import path in
editable installs that are performed via the new PEP660 mechanism used
by default in newer versions of pip and setuptools.

The existing custom `develop` command is retained for compatibility with
older setuptools and pip, which use `python setup.py develop` to create
editable installs. It can be removed once it is safe to assume PEP660
functionality is available and default on all supported pip and
setuptools versions sometime in the future.